изменение типа

+++ src/main/java/BodyMassIndexService.java
@@ -1,7 +1,7 @@
 public class BodyMassIndexService {
-    int calculate(int weight,double heightSm){
+    double calculate(double weight,double heightSm){
         double heightM = heightSm/100;
-        int BMI = (int) (weight/(heightM * heightM));
+        double BMI = weight/(heightM * heightM);
         return BMI;
     }
 }
